pyproject.toml
[project]
name = "personalaigent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atomic-agents>=2.6.0",
    "instructor>=1.13.0",
    "ipykernel>=7.1.0",
    "langfuse>=3.12.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
